Adding support for FreeBSD 9.2
Adding kernel and poudriere support to allow building FreeBSD on 9.2.

This kernel conf is much closer to GENERIC than the previous conf. This means
a larger kernel and less requirement for modules.
Having everything in the kernel is more convenient but could potentially cause
issues, such as when trying to put devices to sleep.