Adding wait_on_cfg.sh: Script to monitor changes to auto save files to cfg
The functionality being added in this commit will allow certain files to
be automatically saved to /cfg when they are changed.

This will hopefully mean that changes made to the most important files are
never accidentally lost and also reduce the overhead of maintenance of the
/cfg as those files will be self-maintained.

This script is initially only monitoring files related to password, users
and groups.

It uses the wait_on script which in turn used kqueue.