Change /var from a memory device to a tmpfs partition
Previously, /var was mounted as a memory device with a predefined size
(around 50MB). In some circumstances this would not be enough and would
result in a full partition. The longer a system was running the more
likely this event would occur.
To resolve this /var is now mounted as a tmpfs partition, giving many
times the previous size (depending on ram size which automatically
dictates the size of the tmpfs partition).

This will also result in more available ram as previously 50MB was
reserved for the partition, regardless of how full the partition was. Now
the partition will only use as much as is stored on it, and due to tmpfs's
ability to also use available swap the possibility of running out of space
on /var has been significantly reduced.

Currently the rc.initdiskless script has been altered and added to the
Files directory. This is not ideal. The changes will be later turned into
a patch and applied during the build process when a patch facility has
been added to the build system.