Renaming the "img_update" script to just "img"
The img_update script has been renamed to just "img" to simplify the name,
making it easier to remember and allows for a broader range of tasks
relating to the image, not just updating.

The "switch" command has also been renamed to "recover" as it more
accurately describes what it does, moves back to the other image. "Switch"
will be reintroduced but with more functionality to handle switching back
and forth better.