create-top.sh: improve install packages function
The function that installed the port packages into the target OS has been
significantly improved.

The function now set uses nullfs mounts to avoid copying packages
unnecessarily and then having to remove them at the end.

It also now sets the location of the package site in the pkg.conf file.
This informs pkg where the package database and package storage directory
are located, in this case as a local url.
This means that packages are now being installed as intended where before
a small hack was used to translate between specified required ports and
pre-built packages in a directory.