create-top.sh: Add special release functions
Adding additional user configurable function type (in addition to customise and late customise)
allowing functions to be created  that are run after all other steps have completed.

This is useful for defining specialist functionality outside of the scripts used specifically in the
release process, once the image has been created.