create-vbox-image.sh: Save result .vdi file to image folder not script folder
Previously the final .vdi image that was created was stored by default in
the same directory that the script was run from.
It seems to make more sense to put it in the same directory as the
original image to avoid polluting your source directory.

To keep simplicity of the script there doesn't seem to be any current need
to allow the user to specify an output directory.